Add original field as fallback
The original field might not be in the fallback chain. `__get__()` falls
back to the original field if all else fails, so it may be surprising
that `instance.field_i18n` falls back to the original field but
`Model.objects.values_list('field_i18n')` does not.

Note that this change might break existing applications.